--- src/sparsify/auto/configs/config_creators/__init__.py
@@ -1,25 +1,0 @@
-# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Submodule containing registered implementations of APIConfigCreator.
-Register decorators will be called when included on initialization of this submodule
-"""
-
-# flake8: noqa
-
-from .helpers import *
-from .image_classification import *
-from .transformers import *
-from .yolov5 import *
