+++ src/sparsify/auto/tasks/image_classification/runner.py
@@ -24,8 +24,7 @@ import torch
 from pydantic import BaseModel
 from sparseml.pytorch.image_classification.export import main as export_hook
 from sparseml.pytorch.image_classification.train import main as train_hook
-from sparsify.auto.api import Metrics
-from sparsify.auto.configs import SparsificationTrainingConfig
+from sparsify.auto.api import Metrics, SparsificationTrainingConfig
 from sparsify.auto.tasks.image_classification.args import ImageClassificationExportArgs
 from sparsify.auto.tasks.runner import DDP_ENABLED, TaskRunner
 from sparsify.auto.utils import HardwareSpecs
