+++ src/sparsify/auto/tasks/transformers/runner.py
@@ -26,8 +26,7 @@ from sparseml.transformers.question_answering import main as question_answering_
 from sparseml.transformers.text_classification import main as text_classification_hook
 from sparseml.transformers.token_classification import main as token_classification_hook
 from sparseml.transformers.utils import SparseAutoModel
-from sparsify.auto.api import Metrics
-from sparsify.auto.configs import SparsificationTrainingConfig
+from sparsify.auto.api import Metrics, SparsificationTrainingConfig
 from sparsify.auto.tasks.runner import TaskRunner
 from sparsify.auto.tasks.transformers import (
     QuestionAnsweringArgs,
