+++ tests/sparsify/auto/autosparse/configs/test_sparse_training_config.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 
-from sparsify.auto import USER_OUT_DIRECTORY
-from sparsify.auto.configs import SparsificationTrainingConfig
+from sparsify.auto import DEFAULT_OUTPUT_DIRECTORY
+from sparsify.auto.api import SparsificationTrainingConfig
 
 
 def test_serialize_sparsificaiton_config_yaml():
@@ -38,7 +38,7 @@ def test_serialize_sparsificaiton_config_yaml():
         task=task,
         base_model=model_path,
         dataset=data_path,
-        save_directory=USER_OUT_DIRECTORY,
+        save_directory=DEFAULT_OUTPUT_DIRECTORY,
         recipe=recipe,
         recipe_args=recipe_args,
         distill_teacher=distill_teacher,
@@ -66,7 +66,7 @@ def test_serialize_sparsificaiton_config_yaml():
     assert (
         sparsification_config_obj.save_directory
         == sparsification_config_reloaded.save_directory
-        == USER_OUT_DIRECTORY
+        == DEFAULT_OUTPUT_DIRECTORY
     )
     assert (
         sparsification_config_obj.recipe
